Refactor: single component in single file
