+++ src/components/app-header/app-header.jsx
@@ -6,7 +6,7 @@ import {
   ProfileIcon,
 } from "@ya.praktikum/react-developer-burger-ui-components";
 import HeaderStyle from "./app-header.module.css";
-import { headerItemPropType } from "../../utils/prop-types.js";
+import HeaderItem from "../header-item/header-item";
 
 class AppHeader extends React.Component {
   constructor(props) {
@@ -61,40 +61,5 @@ class AppHeader extends React.Component {
     );
   }
 }
-
-class HeaderItem extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
-  addInactiveClass() {
-    if (this.props.selected !== this.props.name) {
-      return " text_color_inactive";
-    }
-    return "";
-  }
-
-  render() {
-    const boxClass = `${HeaderStyle.item}` + " pt-4 pb-4 pr-5 pl-5 mt-4 mb-4";
-    const labelClass =
-      `${HeaderStyle.label}` +
-      " text text_type_main-default" +
-      `${this.addInactiveClass()}`;
-    return (
-      <a
-        href="#"
-        className={boxClass}
-        onClick={(e) => {
-          this.props.onClick(this.props.name);
-        }}
-      >
-        {this.props.children}
-        <p className={labelClass}>{this.props.label}</p>
-      </a>
-    );
-  }
-}
-
-HeaderItem.propTypes = headerItemPropType;
 
 export default AppHeader;
