+++ src/components/app-header/app-header.module.css
@@ -13,22 +13,3 @@
   gap: 8px;
   justify-self: stretch;
 }
-
-.item {
-  box-sizing: border-box;
-  margin: 0;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 8px;
-  text-decoration: none;
-  color: unset;
-}
-
-.item:last-of-type {
-  justify-self: end;
-}
-
-.label {
-  margin: 0;
-}
