+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -6,36 +6,3 @@
   flex-direction: column;
   height: 100%;
 }
-
-.card__group {
-  margin: 0;
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  column-gap: 24px;
-  row-gap: 32px;
-}
-
-.card {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  align-items: center;
-}
-
-.price__box {
-  display: flex;
-  flex-direction: row;
-  gap: 8px;
-  align-items: center;
-}
-
-.price {
-  margin: 0;
-}
-
-.label {
-  margin: 0;
-  min-height: 48px;
-  text-align: center;
-}
