+++ src/components/card/card.module.css
@@ -1,0 +1,24 @@
+.card {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+  align-items: center;
+}
+
+.price__box {
+  display: flex;
+  flex-direction: row;
+  gap: 8px;
+  align-items: center;
+}
+
+.price {
+  margin: 0;
+}
+
+.label {
+  margin: 0;
+  min-height: 48px;
+  text-align: center;
+}
