+++ src/components/header-item/header-item.module.css
@@ -1,0 +1,18 @@
+.item {
+  box-sizing: border-box;
+  margin: 0;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 8px;
+  text-decoration: none;
+  color: unset;
+}
+
+.item:last-of-type {
+  justify-self: end;
+}
+
+.label {
+  margin: 0;
+}
