+++ src/components/section/section.module.css
@@ -1,0 +1,7 @@
+.card__group {
+  margin: 0;
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  column-gap: 24px;
+  row-gap: 32px;
+}
